# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OmniChat omnichannel messaging platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Channels
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List all channels
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /channels/{id}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          description: Not found
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: Update a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Delete a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Contacts
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: channelType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  # Conversations
...
  /conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # Webhooks
...
components:
  schemas:
...
    Department:
      type: object
      required: [id, name, routingMode, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        routingMode:
          type: string
          enum: [manual, round_robin]
        isActive:
          type: boolean
        memberCount:
          type: integer
        createdAt:
          type: string
    DepartmentInput:
      type: object
      required: [name, routingMode]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        routingMode:
          type: string
          enum: [manual, round_robin]
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        routingMode:
          type: string
          enum: [manual, round_robin]
        isActive:
          type: boolean

    # Channels
    Channel:
      type: object
      required: [id, name, channelType, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        channelType:
          type: string
          enum: [whatsapp, instagram, facebook]
        externalId:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        pageId:
          type: ["string", "null"]
        accessToken:
          type: ["string", "null"]
        webhookVerifyToken:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    ChannelInput:
      type: object
      required: [name, channelType]
      properties:
        name:
          type: string
        channelType:
          type: string
          enum: [whatsapp, instagram, facebook]
        externalId:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        pageId:
          type: ["string", "null"]
        accessToken:
          type: ["string", "null"]
        webhookVerifyToken:
          type: ["string", "null"]
    ChannelUpdate:
      type: object
      properties:
        name:
          type: string
        externalId:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        pageId:
          type: ["string", "null"]
        accessToken:
          type: ["string", "null"]
        webhookVerifyToken:
          type: ["string", "null"]
        isActive:
          type: boolean

    # Contacts
    Contact:
      type: object
      required: [id, name, channelType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        channelType:
          type: string
          enum: [whatsapp, instagram, facebook]
        externalId:
          type: string
        customFields:
          type: ["string", "null"]
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [name, channelType, externalId]
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        channelType:
          type: string
          enum: [whatsapp, instagram, facebook]
        externalId:
          type: string
        customFields:
          type: ["string", "null"]
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        customFields:
          type: ["string", "null"]

    # Conversations
...
    Message:
      type: object
      required: [id, conversationId, direction, contentType, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderType:
          type: string
          enum: [contact, agent, system]
        senderId:
          type: ["integer", "null"]
        direction:
          type: string
          enum: [inbound, outbound]
        contentType:
          type: string
          enum: [text, image, video, audio, document, location, sticker, template, note]
        content:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        isRead:
          type: boolean
        externalMessageId:
          type: ["string", "null"]
        metadata:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [contentType]
      properties:
        contentType:
          type: string
          enum: [text, image, video, audio, document, location, sticker, template, note]
        content:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        senderId:
          type: ["integer", "null"]

    # Webhooks
...