# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniChat omnichannel messaging platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelChannelType } from './channelChannelType';

export interface Channel {
  id: number;
  name: string;
  channelType: ChannelChannelType;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  pageId?: string | null;
  /** @nullable */
  accessToken?: string | null;
  /** @nullable */
  webhookVerifyToken?: string | null;
  isActive: boolean;
  createdAt: string;
}