# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniChat omnichannel messaging platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInputChannelType } from './contactInputChannelType';

export interface ContactInput {
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  channelType: ContactInputChannelType;
  externalId: string;
  /** @nullable */
  customFields?: string | null;
}