# lib/api-zod/src/generated/types/departmentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniChat omnichannel messaging platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DepartmentUpdateRoutingMode } from './departmentUpdateRoutingMode';

export interface DepartmentUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  routingMode?: DepartmentUpdateRoutingMode;
  isActive?: boolean;
}