# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniChat omnichannel messaging platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageInputContentType } from './messageInputContentType';

export interface MessageInput {
  contentType: MessageInputContentType;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  mediaUrl?: string | null;
  /** @nullable */
  mediaType?: string | null;
  /** @nullable */
  senderId?: number | null;
}